requirements.txt
feedparser==5.2.1
Flask==0.10.1
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
requests==2.10.0
celery==3.1.25
redis==2.10.5
sinaweibopy
